refactor: extract introducer from recommender
